Divide classes into packages
Organize the classes into suitable java packages.
