+++ src/main/java/handsome/Storage.java
@@ -1,9 +1,8 @@
+package handsome;
+
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Scanner;
 
 public class Storage {
     private File handsome;
