+++ src/main/java/handsome/command/DeleteCommand.java
@@ -1,18 +1,25 @@
+package handsome.command;
+
+import handsome.TaskList;
+import handsome.exception.HandsomeException;
+import handsome.Storage;
+import handsome.Ui;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
-    private String input;
+public class DeleteCommand extends Command {
+    private final String input;
 
-    public UnmarkCommand(String input) {
+    public DeleteCommand(String input) {
         this.input = input;
     }
 
     @Override
     public void execute(TaskList tasks, Storage storage, Ui ui) throws HandsomeException, IOException {
         int index = toNumber(input.substring(7), tasks.getSize()) - 1;
-        tasks.markUndone(index);
+        tasks.remove(index);
         storage.writeToFile(tasks);
-        ui.showUnmarkText(index, tasks);
+        ui.showDeleteText(index, tasks);
     };
 
     @Override
