+++ src/main/java/handsome/command/EventCommand.java
@@ -1,3 +1,12 @@
+package handsome.command;
+
+import handsome.TaskList;
+import handsome.exception.HandsomeException;
+import handsome.exception.InvalidTaskException;
+import handsome.Storage;
+import handsome.Ui;
+import handsome.task.Event;
+
 import java.io.IOException;
 import java.util.regex.PatternSyntaxException;
