+++ src/main/java/handsome/command/MarkCommand.java
@@ -1,3 +1,10 @@
+package handsome.command;
+
+import handsome.Storage;
+import handsome.TaskList;
+import handsome.Ui;
+import handsome.exception.HandsomeException;
+
 import java.io.IOException;
 
 public class MarkCommand extends Command {
