+++ src/main/java/handsome/command/PromptCommand.java
@@ -1,3 +1,9 @@
+package handsome.command;
+
+import handsome.Storage;
+import handsome.TaskList;
+import handsome.Ui;
+
 public class PromptCommand extends Command {
     @Override
     public void execute(TaskList tasks, Storage storage, Ui ui) {
