+++ src/main/java/handsome/exception/HandsomeException.java
@@ -1,3 +1,5 @@
+package handsome.exception;
+
 public class HandsomeException extends Exception {
     public HandsomeException(String msg) {
         super(msg);
