+++ src/main/java/handsome/exception/InvalidSyntaxException.java
@@ -1,4 +1,6 @@
-public class InvalidSyntaxException extends HandsomeException{
+package handsome.exception;
+
+public class InvalidSyntaxException extends HandsomeException {
     public InvalidSyntaxException() {
         super("Invalid Syntax! Type prompt to check for valid command");
     }
