+++ src/main/java/handsome/exception/InvalidTaskException.java
@@ -1,3 +1,5 @@
+package handsome.exception;
+
 public class InvalidTaskException extends HandsomeException {
     public InvalidTaskException(String msg) {
         super(msg);
