+++ src/main/java/handsome/exception/ListIndexOutOfBoundException.java
@@ -1,4 +1,6 @@
-public class ListIndexOutOfBoundException extends HandsomeException{
+package handsome.exception;
+
+public class ListIndexOutOfBoundException extends HandsomeException {
     public ListIndexOutOfBoundException(int size) {
         super("Invalid List Index! List has " + size + " items.");
     }
