+++ src/main/java/handsome/task/Event.java
@@ -1,3 +1,7 @@
+package handsome.task;
+
+import handsome.exception.InvalidTaskException;
+
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
